Move Marker into other directory.
